Factor jdbba input validation into shared module and validate autograd entry
